refactored OctData_t

+++ Console/consoleApp/Backend/signalmodel.cpp
@@ -21,14 +21,14 @@ void SignalModel::allocateOctData()
     LOG3(rawDataSize, fftDataSize, dispDataSize); //8192, 4096, 1024, 1982464
 
     for(int i = 0; i < 3; ++i){
-//  for(int i = 0; i < FRAME_BUFFER_SIZE; ++i){
+//    for(int i = 0; i < FRAME_BUFFER_SIZE; ++i){
 
         OCTFile::OctData_t oct;
 
-        oct.advancedViewIfftData  = new quint16 [rawDataSize];
-        oct.advancedViewFftData   = new quint16 [fftDataSize];
-        oct.dispData  = new uchar [dispDataSize];
-        oct.videoData = new uchar [dispDataSize];
+        oct.advancedViewIfftData  = new uint8_t [rawDataSize];
+        oct.advancedViewFftData   = new uint8_t [fftDataSize];
+        oct.dispData  = new uint8_t [dispDataSize];
+        oct.videoData = new uint8_t [dispDataSize];
         oct.acqData   = new uint8_t [MAX_ACQ_IMAGE_SIZE];
 
         m_octData.push_back(oct);
